fix: status code 401 반환하기

+++ backend/src/question-list/question-list.controller.ts
@@ -3,11 +3,13 @@ import {
     Controller,
     Delete,
     Get,
+    HttpStatus,
     Param,
     Patch,
     Post,
     Query,
     Req,
+    Res,
     UseGuards,
     UsePipes,
     ValidationPipe,
@@ -28,31 +30,43 @@ export class QuestionListController {
     constructor(private readonly questionListService: QuestionListService) {}
 
     @Get()
+    @UseGuards(AuthGuard("jwt"))
     @UsePipes(new ValidationPipe({ transform: true }))
-    async getAllQuestionLists(@Query() query: PaginateQueryDto) {
+    async getAllQuestionLists(
+        @Res() res,
+        @Query() query: PaginateQueryDto,
+        @JwtPayload() token: IJwtPayload
+    ) {
         try {
+            if (!token) {
+                // 로그인 안함
+                // isScrap: false
+            }
             const { allQuestionLists, meta } =
                 await this.questionListService.getAllQuestionLists(query);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "All question lists received successfully.",
                 data: {
                     allQuestionLists,
                     meta,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to get all question lists.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Post("category")
+    @UseGuards(AuthGuard("jwt"))
     @UsePipes(new ValidationPipe({ transform: true }))
     async getAllQuestionListsByCategoryName(
+        @Res() res,
+        @JwtPayload() token: IJwtPayload,
         @Query() query: PaginateQueryDto,
         @Body()
         body: {
@@ -60,24 +74,28 @@ export class QuestionListController {
         }
     ) {
         try {
+            if (!token) {
+                // 로그인 안함
+                // isScrap: false
+            }
             const { categoryName } = body;
             query.category = categoryName;
             const { allQuestionLists, meta } =
                 await this.questionListService.getAllQuestionLists(query);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "All question lists received successfully.",
                 data: {
                     allQuestionLists,
                     meta,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to get all question lists.",
                 error: error.message,
-            };
+            });
         }
     }
 
@@ -85,6 +103,7 @@ export class QuestionListController {
     @UseGuards(AuthGuard("jwt"))
     async createQuestionList(
         @JwtPayload() token: IJwtPayload,
+        @Res() res,
         @Req() req,
         @Body()
         body: {
@@ -96,6 +115,11 @@ export class QuestionListController {
     ) {
         try {
             const { title, contents, categoryNames, isPublic } = body;
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
 
             // 질문지 DTO 준비
             const createQuestionListDto: CreateQuestionListDto = {
@@ -109,26 +133,27 @@ export class QuestionListController {
             // 질문지 생성
             const { createdQuestionList, createdQuestions } =
                 await this.questionListService.createQuestionList(createQuestionListDto);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Question list created successfully.",
                 data: {
                     createdQuestionList,
                     createdQuestions,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to create question list.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Post("contents")
     @UseGuards(AuthGuard("jwt"))
     async getQuestionListContents(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Body()
         body: {
@@ -136,62 +161,84 @@ export class QuestionListController {
         }
     ) {
         try {
+            if (!token) {
+                // 로그인 안함
+                // isScrap: false
+                // 비공개 질문 접근 불가 처리
+            }
             const userId = token.userId;
             const { questionListId } = body;
             const questionListContents: QuestionListContentsDto =
                 await this.questionListService.getQuestionListContents(questionListId, userId);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Question list contents received successfully.",
                 data: {
                     questionListContents,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to get question list contents.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Get("my")
     @UseGuards(AuthGuard("jwt"))
     @UsePipes(new ValidationPipe({ transform: true }))
-    async getMyQuestionLists(@Query() query: PaginateQueryDto, @JwtPayload() token: IJwtPayload) {
+    async getMyQuestionLists(
+        @Res() res,
+        @Query() query: PaginateQueryDto,
+        @JwtPayload() token: IJwtPayload
+    ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { myQuestionLists, meta } = await this.questionListService.getMyQuestionLists(
                 userId,
                 query
             );
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "My question lists received successfully.",
                 data: {
                     myQuestionLists,
                     meta,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to get my question lists.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Patch("/:questionListId")
     @UseGuards(AuthGuard("jwt"))
     async updateQuestionList(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Param("questionListId") questionListId: number,
         @Body() body: { title?: string; isPublic?: boolean; categoryNames?: string[] }
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { title, isPublic, categoryNames } = body;
             const updateQuestionListDto: UpdateQuestionListDto = {
                 id: questionListId,
@@ -203,64 +250,78 @@ export class QuestionListController {
 
             const updatedQuestionList =
                 await this.questionListService.updateQuestionList(updateQuestionListDto);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Question list is updated successfully.",
                 data: {
                     questionList: updatedQuestionList,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to update question list.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Delete("/:questionListId")
     @UseGuards(AuthGuard("jwt"))
     async deleteQuestionList(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Param("questionListId") questionListId: number
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const result = await this.questionListService.deleteQuestionList(
                 questionListId,
                 userId
             );
 
             if (result.affected) {
-                return {
+                return res.status(HttpStatus.OK).json({
                     success: true,
                     message: "Question list is deleted successfully.",
-                };
+                });
             } else {
-                return {
-                    success: true,
+                return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
+                    success: false,
                     message: "Failed to delete question list.",
-                };
+                });
             }
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: true,
                 message: "Failed to delete question list.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Post("/:questionListId/question")
     @UseGuards(AuthGuard("jwt"))
     async addQuestion(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Body() body: { content: string },
         @Param("questionListId") questionListId: number
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { content } = body;
             const questionDto: QuestionDto = {
                 content,
@@ -270,31 +331,38 @@ export class QuestionListController {
 
             const result = await this.questionListService.addQuestion(questionDto);
 
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "The new question is added to the list successfully.",
                 data: {
                     questionList: result,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to add the new question to the list.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Patch("/:questionListId/question/:questionId")
     @UseGuards(AuthGuard("jwt"))
     async updateQuestion(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Body() body: { content: string },
         @Param() params: { questionListId: number; questionId: number }
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { content } = body;
             const { questionListId, questionId } = params;
             const questionDto: QuestionDto = {
@@ -306,30 +374,37 @@ export class QuestionListController {
 
             const result = await this.questionListService.updateQuestion(questionDto);
 
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Question is updated successfully.",
                 data: {
                     questionList: result,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to update question.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Delete("/:questionListId/question/:questionId")
     @UseGuards(AuthGuard("jwt"))
     async deleteQuestion(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Param() params: { questionListId: number; questionId: number }
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { questionListId, questionId } = params;
             const deleteQuestionDto: DeleteQuestionDto = {
                 id: questionId,
@@ -338,22 +413,22 @@ export class QuestionListController {
             };
             const result = await this.questionListService.deleteQuestion(deleteQuestionDto);
             if (result.affected) {
-                return {
+                return res.status(HttpStatus.OK).json({
                     success: true,
                     message: "Question is deleted successfully.",
-                };
+                });
             } else {
-                return {
-                    success: true,
+                return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
+                    success: false,
                     message: "Failed to delete question.",
-                };
+                });
             }
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to delete question.",
                 error: error.message,
-            };
+            });
         }
     }
 
@@ -361,90 +436,111 @@ export class QuestionListController {
     @UseGuards(AuthGuard("jwt"))
     @UsePipes(new ValidationPipe({ transform: true }))
     async getScrappedQuestionLists(
+        @Res() res,
         @Query() query: PaginateQueryDto,
         @JwtPayload() token: IJwtPayload
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { scrappedQuestionLists, meta } =
                 await this.questionListService.getScrappedQuestionLists(userId, query);
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Scrapped question lists received successfully.",
                 data: {
                     questionList: scrappedQuestionLists,
                     meta,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to get scrapped question lists.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Post("scrap")
     @UseGuards(AuthGuard("jwt"))
     async scrapQuestionList(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Body() body: { questionListId: number }
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login required.",
+                });
             const userId = token.userId;
+
             const { questionListId } = body;
             const scrappedQuestionList = await this.questionListService.scrapQuestionList(
                 questionListId,
                 userId
             );
 
-            return {
+            return res.status(HttpStatus.OK).json({
                 success: true,
                 message: "Question list is scrapped successfully.",
                 data: {
                     questionList: scrappedQuestionList,
                 },
-            };
+            });
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to scrap question list.",
                 error: error.message,
-            };
+            });
         }
     }
 
     @Delete("scrap/:questionListId")
     @UseGuards(AuthGuard("jwt"))
     async unscrapQuestionList(
+        @Res() res,
         @JwtPayload() token: IJwtPayload,
         @Param("questionListId") questionListId: number
     ) {
         try {
+            if (!token)
+                return res.status(HttpStatus.UNAUTHORIZED).json({
+                    success: false,
+                    message: "Login is required to create question list.",
+                });
             const userId = token.userId;
+
             const unscrappedQuestionList = await this.questionListService.unscrapQuestionList(
                 questionListId,
                 userId
             );
 
             if (unscrappedQuestionList.affected) {
-                return {
+                return res.status(HttpStatus.OK).json({
                     success: true,
                     message: "Question list unscrapped successfully.",
-                };
+                });
             } else {
-                return {
+                return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                     success: false,
                     message: "Failed to unscrap question list.",
-                };
+                });
             }
         } catch (error) {
-            return {
+            return res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
                 success: false,
                 message: "Failed to unscrap question list.",
                 error: error.message,
-            };
+            });
         }
     }
 }
